♻️ Lombok @Builder 적용 및 id 자동 생성 보호
- 기존 수동 Builder 패턴을 Lombok @Builder로 대체하여 코드 간결화
