--- src/main/java/com/burntoburn/easyshift/entity/Token.java
@@ -1,49 +1,0 @@
-package com.burntoburn.easyshift.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Token {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String accessToken;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    protected Token() {
-    }
-
-    private Token(Builder builder) {
-        this.accessToken = builder.accessToken;
-        this.user = builder.user;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public static class Builder {
-        private String accessToken;
-        private User user;
-
-        public Builder accessToken(String accessToken) {
-            this.accessToken = accessToken;
-            return this;
-        }
-
-        public Builder user(User user) {
-            this.user = user;
-            return this;
-        }
-
-        public Token build() {
-            return new Token(this);
-        }
-    }
-}
