+++ src/main/java/com/burntoburn/easyshift/entity/schedule/ScheduleStatus.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.entity;
+package com.burntoburn.easyshift.entity.schedule;
 
 public enum ScheduleStatus {
     PENDING,
