+++ src/main/java/com/burntoburn/easyshift/entity/schedule/Shift.java
@@ -1,0 +1,42 @@
+package com.burntoburn.easyshift.entity.schedule;
+
+import com.burntoburn.easyshift.entity.user.User;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@Getter
+@Entity
+@Table(name = "shift") // 테이블명 명시
+@NoArgsConstructor(access = AccessLevel.PROTECTED) // JPA 기본 생성자 (protected)
+@AllArgsConstructor // 모든 필드를 포함한 생성자 자동 생성
+@Builder // Lombok Builder 적용
+public class Shift {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Setter(AccessLevel.NONE) // ID는 자동 생성되므로 Builder에서 제외
+    private Long id;
+
+    @Column(nullable = false)
+    private String shiftName;
+
+    @Column(nullable = false)
+    private LocalDate shiftDate;
+
+    @Column(nullable = false)
+    private LocalTime startTime;
+
+    @Column(nullable = false)
+    private LocalTime endTime;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "schedule_id", nullable = false)
+    private Schedule schedule;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+}
