+++ src/main/java/com/burntoburn/easyshift/entity/schedule/ShiftTemplate.java
@@ -1,0 +1,33 @@
+package com.burntoburn.easyshift.entity.schedule;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalTime;
+
+@Getter
+@Entity
+@Table(name = "shift_template") // 테이블명 명시
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder // Lombok Builder 적용
+public class ShiftTemplate {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Setter(AccessLevel.NONE)
+    private Long id;
+
+    @Column(nullable = false)
+    private String shiftTemplateName;
+
+    @Column(nullable = false)
+    private LocalTime startTime;
+
+    @Column(nullable = false)
+    private LocalTime endTime;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "schedule_template_id", nullable = false)
+    private ScheduleTemplate scheduleTemplate;
+}
