+++ src/main/java/com/burntoburn/easyshift/entity/user/ApprovalStatus.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.entity;
+package com.burntoburn.easyshift.entity.user;
 
 public enum ApprovalStatus {
     PENDING,
