+++ src/main/java/com/burntoburn/easyshift/entity/user/Role.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.entity;
+package com.burntoburn.easyshift.entity.user;
 
 public enum Role {
     WORKER,
