+++ src/main/java/com/burntoburn/easyshift/entity/user/Token.java
@@ -1,0 +1,25 @@
+package com.burntoburn.easyshift.entity.user;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Entity
+@Table(name = "token") // 테이블명 명시
+@NoArgsConstructor(access = AccessLevel.PROTECTED) // JPA 기본 생성자 (protected)
+@AllArgsConstructor // 모든 필드를 포함한 생성자 자동 생성
+@Builder // Lombok Builder 적용
+public class Token {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Setter(AccessLevel.NONE)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String accessToken;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+}
